linux: fix deconfiguring initramfs compression
Initramfs was disabled when building modules to prevent from an inexisting
rootfs at first build.
But rootfs should be unneeded for modules. Why was it needed ? It had
been observed that modules cannot be built if vmlinux is not build first.
It seems it was a bugged behaviour and "make vmlinux" can be removed. So
no need of disabling initramfs and deconfiguring anymore.